Split Linear issue service by operation
