+++ src/main/runtime/rpc/methods/linear-agent-access.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest'
 import { RpcDispatcher } from '../dispatcher'
 import type { RpcRequest } from '../core'
 import { OrcaRuntimeService } from '../../orca-runtime'
-import { LinearWriteFailure } from '../../../linear/issues'
+import { LinearWriteFailure } from '../../../linear/linear-issue-write-support'
 import { sanitizeLinearErrorMessage } from '../../../linear/issue-context-errors'
 import { LINEAR_AGENT_ACCESS_METHODS } from './linear-agent-access'
